requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
Django==6.0.5
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.15
PyJWT==2.12.1
requests==2.34.2
sqlparse==0.5.5
urllib3==2.7.0
